Updated the file for latest version of all libraries
Updated the script to work on 3.9.13 with all the latest libraries.
Included the code from #1 Issue, the AO code.
Adjusted the code for performance. (Any better suggestions are welcome)
Added search folder for batch conversions.
Added Multiprocessing for the most optimal speed.
Added saving to separate folders for best organization.